fix(ui-ux): fix lint errors
